split doc file
